Add connection setup and initialize streaming
This change will enable the library to connect to the server and use a
streaming replication slot.

The example program is being used for development and debugging
purposes for now until I cut the first release.  After that it should
be a basic test program for users to check that the library can
connect to their server and receive messages, similar to
`pg_receivelogical`.