Add more serialization tests to Types/Lsn.hs
Added a test to check that the ASCII and binary serializers are
equivalent.

The version of `bytestring` in the current LTS we're using doesn't
have upper-case hex encoders.  Postgres encodes in upper case but
hopefully parses insensitive of case.

Upper-case encoders have been added upstream so we should be able to
upgrade the library at some point and use uppercase as well.