Fix the wal sender timeout issue
If we continuosly send a properly formatted `StandbyStatusUpdate`
message with LSNs and all we don't get the timeout issue anymore.

Decided to remove the keep alive response handling code to simplify
the library.

Kept some extension points in place in the `WalCopyData` type and copy
data handlers in case we want to handle those messages in the future.