Add the next LSN to the return type of the callback action
By returning the "next LSN" of the change to replicant we update the
stream state and assume the change has been committed.

Add changeNextLSN to Replicant export list

Also update the haddocks for the `Change` type and `Replicant` module
exports list.

Update README.md

Add information about the changes to the callback and updating the
stream state.