Add ChangePayload
This type parses the two cases of WAL messages we receive:
informational messages emitted by the Postgres function,
`pg_logical_emit_message` (when emitted outside of a transaction), and
change messages containing the usual transaction changes.

We can tell these message types apart because the change messages have
a `nextlsn` key that we need to return to the server.  Informational
messages are purely for information and are not part of the WAL log
stream so they don't have an LSN.

Fixes: #28